requirements_URL_nlu_t2s.txt
Flask
ibm-cloud-sdk-core
ibm-watson
PyAudio
pytest
python-dotenv
